requirements.txt
opencv-python
Pillow
tkinter
face_recognition